+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Next Release
 """""""
 
 - Features:
+    - Add search bar for lookup through state machine libraries
     - Add find usage for finding the usages of state machine libraries